Move some stuff to new utils moduke.
